Add new post script
This is a possible solution for issue #8. Found this script and modified it for us to use and create a new blog post. The user will need to call the script and pass the title of the blog post as shown below and a file will be created in the /content folder with the filename format YYYY_MM_DD_new-post-title.md.

`python new_post.py "New Post Title"`

Inside the file created will be the basic structure for a new markdown post

```
Title: New Post Title
Date: YYYY-MM-DD HH:MM
Modified: 
Author: 
Category: 
Tags: 
```